feat: migrate email push notifiactions and full email sync to pulumi
